Create index.html for blog section
It appears that I did not use the github page with jekyll where it automatically will have a blog section. So I decide to create a blog section on my own. 
Post is designed for a dynamic blog in a webpage. Since the post is made with time, it will show up in the blog section chronologically.